Fixed page reload when new thought was submitted or like button was clicked
